Backend/requirements.txt
autopep8==2.0.2
argon2-cffi==21.3.0
celery==5.2.7
ffprobe==0.5
Flask==2.2.2
Flask-JWT-Extended==4.4.4
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Flask-Cors==3.0.10
pydub==0.25.1
gunicorn==20.1.0
SQLAlchemy==1.4.42
marshmallow-sqlalchemy==0.28.1
redis==4.3.4
Werkzeug==2.2.2
wheel==0.37.1
psycopg2-binary==2.9.4
python-dotenv==0.21.0